requirements.txt
telethon>=1.28.5
aiogram>=3.0.0
aiosqlite>=0.19.0
openai>=1.3.0
requests>=2.28.2
python-dotenv>=1.0.0
httpx>=0.24.1
pydantic>=2.0.0
SQLAlchemy>=2.0.0
asyncpg>=0.27.0
aiohttp>=3.8.4
pytz>=2023.3
websockets>=11.0.3
beautifulsoup4>=4.11.2
matplotlib>=3.7.1
pillow>=10.0.0 